Simulating computing functionality
I have created a C program that contains a simulation of error messages, ad messages, emailing someone and a password login terminal. This program also contains a custom header file, and a custom ASCII input converter (hidden password). If you are not using a Linux machine, I have a .mp4 file included in order for you to view the functionality.